Fix build for pg 11devel.
Now that feature freeze is in effect, fix the build to account for
missing TRUE/FALSE and the type change to pq_sendbyte.

Not seeing an immediate need for a release; I'll do that sometime
during beta/rc for 11.0.

+++ ipaddr.c
@@ -266,7 +266,7 @@ ipaddr_send(PG_FUNCTION_ARGS)
 
     pq_begintypsend(&buf);
 	pq_sendbyte(&buf, af);
-	pq_sendbyte(&buf, ipr_af_maxbits(af));
+	pq_sendbyte(&buf, (int8) ipr_af_maxbits(af));
 	pq_sendbyte(&buf, 1);
 	pq_sendbyte(&buf, ip_sizeof(af));
